Move PwsFile data to application class
The PasswdSafeApp application class now holds the open PwsFile and its parsed data.  Now, the data can be reused when the PasswdSafe activity is restarted for the same file due to screen rotation, etc.  Also, the data can now be reused with the RecordView activity to view a particular record.